New exercise: MEDIEVAL CANNON SIMULATOR
Added a new program that computes the trajectory of a granite projectile with and without the presence of a drag force.

This exercise shows an implementation of the explicit Euler method and the Runge Kutta 4 method, with an ideal case study of the precision for each method.

Right now it lacks optimization, although the precision study is inherently slow.