Added descriptions to the dobule pendulum problem
Added the assignement of the problem, with a description and tasks to complete.

In the program, the time of simulation has been changed to 60s (it should take about 15s to compute for 10 pendulums).

Also changed the animation interval to match the real time simulation and added a brief docstring to the propagate routine.