Added support for uniform dielectric materials and linear capacitance computation
Added an uniform dielectric material between the two plates and updated the simulation loop to use that data to compute the laplacian.

Removed the electric field plot, which shows reasonable difficulty to express the reality since the simulation cell is finite.

Added the linear capacity computation, which differs from the reality, probably because of border effects.

Also added the assignement to the problem.